Add files via upload
"printerror" header file has been moved to the main file for clarity.

Bug in the outputs where the program would print empty response messages has been fixed.

Program now takes the number of columns from environment variables and prints the coming responses at a fixed distance from left. Though, this causes a problem when the terminal size is changed, where the the program prints the messages at the same distance from the left. Additionally, if the coming message is too long the rest of the message is printed from the beginning from left, for UI purposes, this should be fixed.

If a large text is copy/pasted, program prints the text until a newline is found. The user has to enter a newline again to send the rest. 

Program also needs to handle EOF and close the program. This feature should be easy to implement.